Añadido el cálculo de coordenadas del tablero del Picross (sin offset). Corregida la forma de crear la matriz, estaban cambiadas las filas y las columnas... Corregido renombrado de clases.

+++ Classes/Picross.cpp
@@ -4,7 +4,7 @@
 USING_NS_CC;
 using namespace std;
 
-Picross::Picross(int num, GameMode gm)
+Picross::Picross(short num, GameMode gm)
 {
 	string mode = "";
 	if(gm == GameMode::NORMAL)
